--- go/vt/vttablet/tabletserver/wasm/common/types.go
@@ -1,1 +1,0 @@
-package common
