+++ go/vt/vttablet/tabletserver/wasm_plugin_wasmer.go
@@ -4,7 +4,6 @@ import (
 	"encoding/binary"
 	"encoding/json"
 	"sync"
-
 	"vitess.io/vitess/go/vt/log"
 
 	"github.com/wasmerio/wasmer-go/wasmer"
